New 1.2 version. Exclude candidates ending in 5.
This is the first instance of the versión 1.2

In this version the candidates 2, 3 and 5 are treated separatedly and then a new main loop
is started where only candidates ending in 7, 9, 1 and 3 are considered, the ones ending
in 5 are skipped, saving some 20% of time.

The tests for this program are executed in the levels 7, 8, 9 and 10, the solving times
are reduced but not enough to beat level 10.