Compute max candidate only at the begining of factorize and after a factor is found
Instead of updating the max candidate for every candidate tried, this time it is computed
at the begining of the factorize function and when a new factor is found.  This reduces
greatly the number of computations made in the function.
The computation is a bit baroque:
    int(math.ceil(math.sqrt(compnum)))
First the square root is found, if it is not an integer (most likely case) we get the next
integer, but ceil retunrs a float of the form 23.0, so the int() funtion is called at the
end.