Correct the bandly broken factorize function that uses a initial candidate
The new factorize function that uses a non default initial candidate, has been fixed from
a non function state.

It now considers the different increment to apply depending on the final figure of the
candidate.

A dictionary of list is used to keep the correct increments from one candidate to the
next.

The inner workings of the function are explained in the README file: FACTORIZACION POR SECTORES