Simplify the computation of the max_candidate when using a top candidate factor
In the function factorize_with_limits the logic to compute the top candidate to look for
factors has been changed, now we use the internal python function
min(last_candidate,[val of max_candidate]).  This removes the somehow weird use of an if
clause, and the separate computation of max_candidate.