Use multiprocessing module to paralellize the factoring of numbers
To improve the use of the hardware, the factoring of the segments is done in a separate
process, as many as CPUs has the computer.  This improves the detection of prime numbers,
but slows the factoring of numbers made up of multiple factors.

The independent processes are based on the factorize_with_limits function, that has been
slightly modified for this purpose

To collect the results a multiprocessing.queue is used.