WIP. Update global max candidate when first process finds a factor.
This version is not complete.

The idea is to have a common variable between the factorize processes to hold the maximum
candidate value, and update it when the lowest factorize process alive finds a new factor.
There are a few probles to have it working, but the formost right now is to signal de
upstream factorize processes that the max candidate has been updated so they can see if
they have to update the local max candidates.
May be we can use an event that gets checked when the local max candidate is checked.  I
don't know how much this will slow the program, a lot probably.