When a process end terminate the rest if factoring is done
Since we wait for the completion of the processes in order, when one ends we check for the
status of the variable mis_acomplish, if it is True, the factoring of the original number
is done and the rest of the processes are terminated.

Now we use a Manager to share variables between the factoring processes and the factor
broker, each factor process (factorize_with_limits) has its own shared list of factors
found and its own Namespace wich currently holds the mis_acomplish boolean variable.

The previous Queue to store the results has been removed since now it is important to
check the results in order, and we couldn't garantee any order before, since the processes
can end a different order from that used to launch them.