Use a Manager object to share the global max candidate among processes
With the help of a Manager object tow pieces of information are shared among the
factoring processes: the name of the first process and the global max candidate, which is
the max candidate of the number to factor disregarding the segment limits.

When a factor is found in the first factoring process the global max candidate is
updated, and can be read by the other processes, which in turn can update their local max
candidates.

Unfortunately the update of the local max candidate only happends when a factor is found
in the process, so if no factor is found the total factoring time is not reduced.