New command line parameter --segments. Keep the found factors on relaunched processes
A command line option (--segments) has been added, it's used to define the number of
segments and their limits.  Use an even number of integers to define the segments.  No
validity check of these number is performed so it's up to the user to make sure they are
sane.  The advantage of this option is that the limits don't have to be evenly spaced,
and this makes it easier to check different use case in the program.

There was a bug in the relaunching of processes, where the previously found factors were
discarded, now they are kept by reusing the own_factors variable from the old process.